+++ tests/Hydrator/NestedMapping/Nested.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\Hydrator\Tests\Hydrator\NestedMapping;
+
+final class Nested
+{
+    public string $var = '';
+    public ?Nested2 $nested2 = null;
+}